proc_forbid(): extensively document why we don't need locking
Also change the ASSERT() in proc_permit() to check for != 0, as a
workaround for CPUs like the AVR which are unable to write an int to
memory atomically.

A better fix would be changing the type of preempt_forbid_cnt to
cpu_stack_t or a new type cpu_atomic_t.