list: check against double node insertion
Tighten debug consistency checks in ADDHEAD(), ADDTAIL() and ENQUEUE()
through the macro LIST_ASSERT_NOT_CONTAINS().  Suggestions for a
better name very welcome.

This check is expensive (O(n)) and too permissive, as what we should
really check for is the node not being already part of *any* list.

This could be simplified by changing the list API to demand
pre-initialization of the node pointers to NULL.