lm3s1968: merge the DevelGPS example application.
The purpose of this project is to implement a portable GPS locator / geocaching
finder tool using the Cortex-M3 based Luminary Micro LM3S1968 evaluation board.

The device proposes a user interface to configure the GPS coordinates of a
target location and shows on the OLED display the direction and distance to
reach the target.