ARM7TDMI: optimize IRQ macros.
Instead of using register "r0" explicitly in IRQ_ENABLE(),
IRQ_DISABLE(), etc. consider to let the compiler decide the best
register to use.

Moreover, add "memory" to the clobber list so that all the IRQ functions
implicitly add a memory barrier. In this way we're sure memory operation
ordering is always preserved on either side of the IRQ operation.