CodeLite: add a small delay during the debugger initialization.
Introduce a small delay (1 sec) before attaching GDB to the debugging
socket.

At the moment this is a requirement only for the lm3s boards, because
openocd claims that the connection for GDB is ready even when it is not
properly initialized.

To a large degree, it is safer to add such small delay also for the
other boards to be sure that the debugger can reliably connect to the
debugging socket.