Add unbuffered methods for accessing a KBlock.
Since some modules which will use KBlock use an
internal buffer, the buffered API supplied by KBlock
may be redundant.

Specifically, if a module (like the FatFS) has its
own internal block buffer, the buffer inside the KBlock
would be wasted.
In order to avoid this waste of RAM, this patch adds
an unbuffered and simplified version of the API.

All KBlocks will implement this unbuffered API, and
they will give the user the option of adding a software
emulated buffer at init time. In this way we can have both
APIs with no RAM wasting.

In order to simplify KBlock implementors' job, a series of 
generic functions is supplied. If the device supports
only a few functions of the KBlock low level API, the missing
ones may be generated using the generic ones.