lm3s: fix build bug in empty and kernel presets (lm3s_ssiOpen)
From r3911 the LM3S SSI driver is not a distinct driver, and it is
supposed to be explicitly included by the user, at least until the
refactoring of the generic serial/spi/ssi module.

This causes build bugs in empty and kernel presets, triggering the
error: undefined reference to `lm3s_ssiOpen'.

As a temporary workaround, explicitly include the LM3S SSI driver into
the user's makefile for empty and kernel presets.

Signed-off-by: Andrea Righi <arighi@develer.com>